leaf81: modularize tokenizer internals
